backend/requirements.txt
APScheduler==3.11.2
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
et_xmlfile==2.0.0
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-Cors==4.0.0
fsspec==2026.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.1
openpyxl==3.1.2
pandas==2.3.3

pymongo==4.6.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pytz==2025.2
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
Werkzeug==3.1.5
wheel==0.45.1
requests

